Moved view related filter code into the filters directory
* Renamed ActiveAdmin::FilterFormBuilder
  to ActiveAdmin::Filters::FormBuilder
* Renamed ActiveAdmin::ViewHelpers::FilterFormHelper
  to ActiveAdmin::Filters::ViewHelper
