+++ lib/active_admin/filters.rb
@@ -1,10 +1,8 @@
-ActiveAdmin.before_load do |app|
+require 'active_admin/filters/dsl'
+require "active_admin/filters/resource_extension"
+require 'active_admin/filters/forms'
 
-  require 'active_admin/filters/dsl'
-  require "active_admin/filters/resource_extension"
-
-  # Add our Extensions
-  ActiveAdmin::ResourceDSL.send :include, ActiveAdmin::Filters::DSL
-  ActiveAdmin::Resource.send :include, ActiveAdmin::Filters::ResourceExtension
-
-end
+# Add our Extensions
+ActiveAdmin::ResourceDSL.send :include, ActiveAdmin::Filters::DSL
+ActiveAdmin::Resource.send    :include, ActiveAdmin::Filters::ResourceExtension
+ActiveAdmin::ViewHelpers.send :include, ActiveAdmin::Filters::ViewHelper
